So it finally connects...
Important point:

I ported the ip address of the webservice to a port in my phone that
would be accessible to the app.

Then I finally got around that INTERNET permission I needed haha (even
though I did not know I needed it at the time).

And, now it only loads one item in the whole view, but I will solve that
as soon as I lay my hands on it